requirements.txt
fastapi
uvicorn
psycopg2-binary
SQLAlchemy
python-dotenv
requests
requests-oauthlib
pydantic
python-multipart
scikit-learn
pandas
numpy
joblib
aiofiles
